+++ front/src/actions/movie.ts
@@ -16,3 +16,6 @@ export const LOAD_SINGLE_MOVIE_FAIL="LOAD_SINGLE_MOVIE_FAIL";
 export const LOAD_RELATED_SEARCH_REQUEST="LOAD_RELATED_SEARCH";
 export const LOAD_RELATED_SEARCH_SUCCESS="LOAD_RELATED_SEARCH_SUCCESS";
 export const LOAD_RELATED_SEARCH_FAIL="LOAD_RELATED_SEARCH_FAIL";
+
+//영화 저장
+export const SAVE_MOVIE="SAVE_MOVIE";
