+++ front/src/components/Movie/MovieCard/index.tsx
@@ -1,6 +1,8 @@
 import React , { useCallback } from 'react';
 import Router from 'next/router';
+import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
+import { SAVE_MOVIE } from '../../../actions/movie';
 import { MovieList } from '../../../model/MovieList';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons';
 import { titleParser } from '../../../util/titleParser';
@@ -13,9 +15,14 @@ interface Props {
 
 // 검색창 검색시 해당 영화 리스트 가져오기
 const MovieCard=({Movie, Search=false}:Props)=>{
+    const dispatch = useDispatch();
 
-    const onSelectMovie=useCallback(()=>{ 
-        Router.push(`/writeReview/${Movie.title}`); // redirect
+    const onSelectMovie=useCallback(()=>{  // 리뷰 작성할 영화 선택 
+        dispatch({
+            type:SAVE_MOVIE,
+            data:Movie,
+        });
+        Router.push(`/writeReview`); // redirect
     },[]);
 
     return(
